convertFortAmount() Will now Return String
convertFortAmount() Amount will now Send the amount value as String, and The same will be used for String Calculation. 
This will resolve the the php bug(7.1.x), which adds Floating point decimal to Integer Values. 
https://bugs.php.net/bug.php?id=72567

The Fort accept Both Integer and String type for Amount, hence it will not impact The Transaction.